added StatusRepository
